requirements.txt
poke-env>=0.8.1
stable-baselines3>=2.3.0
torch>=2.2.0
gymnasium>=0.29.0
numpy>=1.26.0
pyyaml>=6.0
tensorboard>=2.17.0
